currentID

+++ Chapssal_Project/src/main/java/com/chapssal/message/repository/ParticipantRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ParticipantRepository extends JpaRepository<Participant, Long> {
+public interface ParticipantRepository extends JpaRepository<Participant, Integer> {
     List<Participant> findByRoom_RoomNum(int roomNum);
 }
